Fix typo in word thermostat
